pyproject.toml
[tool.poetry]
name = "notenook-temlate"
version = "0.1.0"
description = ""
authors = ["ignition <valentijn.ignition@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.2"
matplotlib = "^3.8.2"
pillow = "^10.2.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
